dev2/workspace/python/requirements.txt
hypothesis
pytest
neo4j
psycopg2-binary
redis
pymemcache
pymysql
